Protect API endpoint using JWT
New classes to generate and pass JSON Web Token to front end where it is stored in the local storage of the browser.  Order component was updated to require this token for the create method.

Normally this would be set up differently so code does not have to be duplicated in several areas.  This is set up for demonstration purposes only.